third_party/blink/web_tests/platform/generic/fast/animation/request-animation-frame-missing-arguments-expected.txt

Test how animation methods react to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS requestAnimationFrame() threw exception TypeError: Failed to execute 'requestAnimationFrame' on 'Window': 1 argument required, but only 0 present..
PASS cancelAnimationFrame() threw exception TypeError: Failed to execute 'cancelAnimationFrame' on 'Window': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
